moved event processing to own package
